Update and rename .java to Minimum times A has to be repeated such that B is a substring of it.java
